When STARPU_SCHED is specified, only execute tests for this scheduler
(cherry picked from commit ed3fa8d2831ff9067cc6bde7aba3ab9c0c4492ec)

+++ tests/perf_models/error_model.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 # StarPU --- Runtime system for heterogeneous multicore architectures.
 #
-# Copyright (C) 2009-2011,2014                           Université de Bordeaux
+# Copyright (C) 2009-2011,2014, 2019                           Université de Bordeaux
 # Copyright (C) 2010,2015,2017                           CNRS
 #
 # StarPU is free software; you can redistribute it and/or modify
@@ -24,6 +24,9 @@ niter=5
 
 #nblockslist="4 4"
 #niter=2
+
+# Testing another specific scheduler, no need to run this
+[ -z "$STARPU_SCHED" -a "$STARPU_SCHED" != dm ] || exit 77
 
 
 rm -f log
